Rework the flow of downloading media
The progress window now acts as the setup window, only showing when needed.

When downloading/unzipping media, the same window is reused across the entire process, instead of being destroyed and remade for each step. This means the window must be manually destroyed when the download is finished (see the settings.ts change).

The functions in download.ts/unzip.ts have been merged in with media.ts.

Also, the popups for first-time startup/mod errors/serverless error are a part of the `mainWindow`, shown after it has been created, since its not necessary to show them beforehand.
