+++ src/client/views/progress/progress.ts
@@ -1,7 +1,7 @@
 import path from 'path'
 import { app, BrowserWindow, dialog } from "electron";
 
-function createProgressBarWindow(prompt: string) {
+export async function createProgressBarWindow() {
   const progressBarWindow = new BrowserWindow({
       width: 300,
       height: 200,
@@ -36,13 +36,15 @@ function createProgressBarWindow(prompt: string) {
     }
   })
 
-  progressBarWindow.loadFile(path.join(__dirname, 'progress.html'));
-  
-  progressBarWindow.on('ready-to-show', () => {
-    progressBarWindow.webContents.send('prompt-name', prompt)
-  })
+  await progressBarWindow.loadFile(path.join(__dirname, 'progress.html'));
 
   return progressBarWindow
+}
+
+export function setPrompt(prompt: string, window: BrowserWindow) {
+  if (window && !window.isDestroyed()) {
+    window.webContents.send('prompt-name', prompt);
+  }
 }
 
 function setProgress(value: number, window: BrowserWindow) {
@@ -53,8 +55,7 @@ function setProgress(value: number, window: BrowserWindow) {
 
 export type ProgressCallback = (progress: number) => void
 
-export async function showProgress(message: string, task: (progress: ProgressCallback, end: () => void) => Promise<boolean>): Promise<boolean> {
-  const progressWin = createProgressBarWindow(message)
+export async function showProgress(progressWin: BrowserWindow, task: (progress: ProgressCallback, end: () => void) => Promise<boolean>): Promise<boolean> {
   let lastUpdate = Date.now();
   return await task((progress: number) => {
     const cur = Date.now();
@@ -62,8 +63,5 @@ export async function showProgress(message: string, task: (progress: ProgressCal
       lastUpdate = cur
       setProgress(progress, progressWin)
     }
-  }, () => {
-    // destroy, not close, otherwise we'll trigger the popup in the 'close' event
-    progressWin.destroy()
-  })
+  }, () => {})
 }
