--- src/client/views/setup.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link href="base.css" rel="stylesheet" type="text/css" media="screen"/>
-  <style>
-    html {
-      height: 100%;
-      display: flex;
-      justify-content: center;
-    }
-    
-    body div {
-      font-size: small;
-      height: 100%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-  </style>
-</head>
-<body>
-  <div>
-    Waddle Forever is starting...
-  </div>
-</body>
-</html>
